Move observers implementations into PeerConnection.
